Added clang-format file based on industry standards and our practices
`clang-format`, like `pep8` for python, is the de-facto standard for C++.
Further, it is pre-installed on most computers, and comes with clang. Thus, it
is the most widely accessible. Thus, I suggest we make it a rule for everyone to
use it, rather than having different individuals use different beautifiers. The
one caveat is that the block indent feature is only available on a newer build.
Thus, although functions arguments will be broken into new lines, the closing
parenthesis will not be on its own line.