Remove compiler selection code from CMakeLists.txt
Turns out, we were using the default compiler the entire time as we set the compiler too late in CMakeLists.txt. As soon as we set up the project with  `project(cartogram LANGUAGES CXX)`, the default compiler was selected (Apple's clang on macOS) and the compiler selection code was never executed.